+++ include/menu.h
@@ -1,0 +1,9 @@
+#ifndef MYSTER_US_ADVENTURES_MENU_H
+#define MYSTER_US_ADVENTURES_MENU_H
+#include <stdio.h>
+
+void open_scores();
+void title_screen();
+
+
+#endif //MYSTER_US_ADVENTURES_MENU_H
