+++ usual.c
@@ -10,6 +10,7 @@
 #include <errno.h>
 
 #include "include/macro.h"
+#include "include/struct.h"
 
 // efface tout le terminal visible
 void clear_all(){
@@ -26,12 +27,6 @@ void clear_part(int line, int column){
 void cursor_move(char direction, int num){
     printf("\033[%d%c", num, direction);
 }
-/*
-// Va vider le "buffer" pour éviter les fuites de donnée quand on fait des getchar notamment
-void flush_input_buffer() {
-    int c;
-    while ((c = getchar()) != '\n' && c != EOF);
-}*/
 
 /* 0: success -1: error */
 int setBlockingFD(int fileDescriptor, int blocking){
@@ -103,17 +98,32 @@ void write_crash_report(const char* error_message) {
     printf("Le rapport de crash a été créé : %s\n", filename);
 }
 
-//permet de print le contenu d'un fichier dans le terminal. Retourne 1 si l'affichage s'est bien passé, 0 sinon.
-int printFile(char filename[]){
-    FILE *file = fopen(filename, "r");
-    if (file == NULL){
-        return 0;
+// Cherche l'indice du symbole de la case dans un tableau spécifié, renvoie l'indice de la première occurrence du symbole dans le tableau, sinon renvoie -1
+int SymbolIdInArray(Square square, const Entity array[], int size){
+    for(int i = 0; i<size; i++){
+        if(strcmp(square.symbol.name, array[i].name) == 0)
+            return i; // retourne le premier indice correspondant si trouvé dans le tableau
     }
+    return -1; //si pas dans le tableau
+}
 
-    char character;
-    while ((character=fgetc(file)) != EOF){
-        printf ("%c", character);
+void background(int i, int j) {
+    for (int k = 0; k < i; k++) {
+        for (int l = 0; l < j; l++) {
+            printf("%s ", B_BLK);
+        }
+        printf("\n");
     }
-    fclose(file);
-    return 1;
+    cursor_move('A', 10);
+    clear_all();
+}
+
+//Vérifie si la case entrée est déjà retournée, ou est le bord de la map, ou est une case de départ. Renvoie 1 si elle est l'une de ces conditions, 0 sinon.
+int CheckSquareInvalid(Square square) {
+    return (square.flipped == 1 || strcmp(square.symbol.name, " ") == 0 || strcmp(square.symbol.name, START) == 0);
+}
+
+//Vérifie si l'index spécifié est valide pour être recherché dans un tableau. Renvoie 1 si hors du tableau, 0 sinon.
+int CheckIndexOutOfArray(int i, int arraySize){
+    return (i < 0 || i >= arraySize);
 }
